refactor: split gateway acp read models
